Move WithScope & WithRedirectUri to different types file
